Safari-609.1.11.1/LayoutTests/canvas/philip/tests/2d.pattern.animated.gif.html

<!DOCTYPE html>
<title>Canvas test: 2d.pattern.animated.gif</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

deferTest();
setTimeout(function () {
    var pattern = ctx.createPattern(document.getElementById('anim-gr.gif'), 'repeat');
    ctx.fillStyle = pattern;
    ctx.fillRect(0, 0, 50, 50);
    setTimeout(wrapFunction(function () {
        ctx.fillRect(50, 0, 50, 50);
        _assertPixelApprox(canvas, 25,25, 0,255,0,255, "25,25", "0,255,0,255", 2);
        _assertPixelApprox(canvas, 75,25, 0,255,0,255, "75,25", "0,255,0,255", 2);
    }), 250);
}, 250);


});
</script>
<img src="../images/anim-gr.gif" id="anim-gr.gif" class="resource">
